requirements.txt
Django==3.0.4
djangorestframework==3.10.3
pytz==2019.2
sqlparse==0.3.0
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
django-rest-auth==0.9.5
django-money==1.0
django-cors-headers==3.2.1
django-import-export==2.1.0
django-extensions==2.2.9
pandas==1.0.3
numpy==1.18